app: Auto reformate code import

+++ app/src/main/java/com/kabouzeid/gramophone/helper/StackBlur.java
@@ -25,12 +25,12 @@ import java.util.concurrent.Executors;
  * on the left side of the stack.
  *
  * @author Enrique López Mañas <eenriquelopez@gmail.com>
- *         http://www.neo-tech.es
- *         <p/>
- *         Author of the original algorithm: Mario Klingemann <mario.quasimondo.com>
- *         <p/>
- *         Based heavily on http://vitiy.info/Code/stackblur.cpp
- *         See http://vitiy.info/stackblur-algorithm-multi-threaded-blur-for-cpp/
+ * http://www.neo-tech.es
+ * <p/>
+ * Author of the original algorithm: Mario Klingemann <mario.quasimondo.com>
+ * <p/>
+ * Based heavily on http://vitiy.info/Code/stackblur.cpp
+ * See http://vitiy.info/stackblur-algorithm-multi-threaded-blur-for-cpp/
  * @copyright: Enrique López Mañas
  * @license: Apache License 2.0
  */
@@ -153,9 +153,9 @@ public class StackBlur {
                     if (i <= wm) src_i += 1;
                     stack_i = i + radius;
                     stack[stack_i] = src[src_i];
-                    sum_r += ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
-                    sum_g += ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
-                    sum_b += (src[src_i] & 0xff) * (radius + 1 - i);
+                    sum_r += (long) ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
+                    sum_g += (long) ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
+                    sum_b += (long) (src[src_i] & 0xff) * (radius + 1 - i);
                     sum_in_r += ((src[src_i] >>> 16) & 0xff);
                     sum_in_g += ((src[src_i] >>> 8) & 0xff);
                     sum_in_b += (src[src_i] & 0xff);
@@ -242,9 +242,9 @@ public class StackBlur {
 
                     stack_i = i + radius;
                     stack[stack_i] = src[src_i];
-                    sum_r += ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
-                    sum_g += ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
-                    sum_b += (src[src_i] & 0xff) * (radius + 1 - i);
+                    sum_r += (long) ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
+                    sum_g += (long) ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
+                    sum_b += (long) (src[src_i] & 0xff) * (radius + 1 - i);
                     sum_in_r += ((src[src_i] >>> 16) & 0xff);
                     sum_in_g += ((src[src_i] >>> 8) & 0xff);
                     sum_in_b += (src[src_i] & 0xff);
